plugin: use configured weights in priority calc
Instead of using hard-coded values for the integer weights for each factor
in the multi-factor priority calculation, use the values located in the
priority_weights map that extract the weights during a conf.update.

If no values are found in a conf.update, the values for the factors located in
the priority_weights map will be -1, so in this case, just use default values
so that a job does not get held in PRIORITY state.