plugin: use configured weights in priority calc
Problem: The priority_calculation () function uses hard-coded values for
its calculation of a job's priority, but the plugin can read from a
configurable TOML file now via the conf.update callback.

Use the values located in the priority_weights map.

If no values are found in a conf.update, the values for the factors
located in the priority_weights map will be -1, so in this case, just
use default values.