plugin: add default bank name to jobspec
Problem: If an association submits a job under a default bank, there is
no record of what the default bank is in the jobspec or eventlog of the
job.

Add the bank name via jobspec-update under attributes.system.bank
during job.new in the case where a user is submitting a job under
their default bank.

In the case where a user submits a job before any flux-accounting info
is sent to the plugin, their job entry will be associated with the
special "DNE" bank until their information is updated.

In this case, once the information is updated and the submitted
job is once again going through job.state.priority, add the
jobspec-update call inside the if-else branch where the bank_info
struct for the job is being updated.