testsuite: stop all queues with --all option
Problem: flux-core PR 4776 added the ability for queues to be
individually started and stopped.  As an update, starting/stopping
all queues with the `flux queue` command now requires the --all option
to be specified, otherwise a warning is output.

To avoid the warning, update all callers to specify --all when multiple
queues are started or stopped.