cmd: add cmd to run accounting service as process
Problem: flux-account commands can currently only be executed on the
management node of a cluster (where the flux-accounting DB is
installed). So, users/admins who want to run flux-account commands to
see user or bank information cannot run these commands on other nodes
on a cluster. There is a need to enable commands to interface with the
DB from other nodes.

Add a new service to flux-accounting that registers with the broker to
accept flux-account commands in the form of RPCs. This will be used in
conjunction with the front-end flux-accounting script, flux-account.py,
to accept flux-accounting commands in the form of RPCs instead of
direct calls to the Python bindings.