queue_subcommands: replace print() statements
Problem: The Python functions in the Accounting class directly emit output to
STDOUT instead of returning information to the caller of the function. To
prepare for the transition to using a systemd-managed service that accepts an
RPC to call the Python functions, this behavior should be changed to return
information to the caller of the function.

Change the behavior of the bindings that emit output to return any potential
output instead of print()-ing it to STDOUT. It will now be the responsibility of
the caller of the function to call the appropriate function, fetch and unpack
any potential data, and print it out.