edit-queue: add int variable type to optional args
Problem: The edit-queue commands currently takes any optional arguments to
the edit-queue command as strings when calling the edit_queue() function. All
of the optional arguments are expected to be integers, so they should be
declared as such, and the command should fail if the wrong type is input on the
command line.

Add an int variable type to all of the optional arguments of the edit-queue
command.