flux-account.py: add return_val for bindings
Now that the Python functions have changed to return any information to
be emitted to STDOUT instead of print()-ed directly, add a return value
to assign any output to in flux-account.py. Then, in this front-end script,
print any output that is not a 0 return code (indicating a successful run).