bank_subcommands: remove -t option for view-bank
Problem: Since we are moving to return information for the functions that
currently print info to STDOUT, we need to adjust some of the functions
slightly. The "view-bank" command is one of these, as the "-t" option will
need to be reworked. In order to make progress on more important fronts, we
should just remove this option for now.

Remove the "-t" flag from the "view-bank" command, which prints out any sub
banks under the bank specified when calling the "view-bank" command. Similar
functionality already exists in the "flux account-shares" command. As a
result, also remove the print_sub_banks() helper function from
bank_subcommands.py. Remove any sharness tests in t1007-flux-account.t that
call the view-bank command with the "-t" flag.