flux-account.py: remove connection attempt to DB
Problem: Now that the flux-accounting service is responsible for
connecting to the flux-accounting database, flux-account.py does not
need to connect to the DB before sending an RPC to the service (with
the exception of the create-db command).

Remove the attempts to connect to the flux-accounting DB in
flux-account.py. Remove the now unused Python modules that were
imported.