flux-account.py: update flux-account commands
Problem: The commands in flux-account.py issue direct calls to the
Python bindings in flux-accounting when they should be issuing RPCs to
the new flux-accounting service, flux-account-service.py.

Change the flux-account commands to issue an RPC to the new Python
flux-accounting service instead of directly calling the appropriate
Python binding.