pylint: disable some pylint checks
Problem: pylint complains about the following as a result of adding
the new flux-accounting service:

too-many-statements: select_accounting_function() now has more than
the allowed number of lines according to pylint by 6 statements.

Disable this check.

broad-except: flux-account-service.py checks for a general Python
exception in a number of its functions because we want to avoid
exiting unexepectedly if there was an error while handling an RPC
request. This is discouraged by pylint. However, we are logging the
exception.

Disable this check.