bank_subcommands: improve context of exceptions
Problem: The add_bank() and edit_bank() functions do not provide clear
error messages on ValueError exceptions which can make it hard to
decipher what error occurred.

Improve the error handling of both of these functions by adding more
context to the message being returned in the case of a ValueError.

Add a sharness test case to t1007-flux-account.t to make sure the
appropriate error message is emitted on an error.