flux-account.py: use formatter class for --help
Problem: flux-account.py does not use the same formatting class as the
Python commands in flux-core, which creates inconsistencies in the way
help output for these commands is presented.

Use the same formatting class as flux-core to make the help output of
commands appear more consistent with other Flux utilities.