service: fix call to output_job_records
Problem 1: The call to output_job_records() in flux-account-service.py
currently fails due to a mismatch in arguments to the function. This
is because the parameters in the call to output_job_records() are not
labeled.

Problem 2: The path passed to the output_job_records() function is
incorrect. The path should be to the flux-core job-archive DB, not
the flux-accounting DB.

Label the parameters in the output_job_records() function.

Adjust the DB path passed to the output_job_records() function to be
the flux-core job-archive DB.