job_archive_interface: adjust fetch_job_records()
Problem: The functions in job_archive_interface.py that construct
strings of job records to both be written to an output file or to be
printed to STDOUT were never adjusted when flux account-service was
added. As such, their output when calling the view-job-records command
resulted in wonky output.

Adjust the fetch_job_records() helper function and the
output_job_records() function to account for the new flux-accounting
service so that output is correct.