edit-user: make userid an editable field
Problem: The "userid" field for a user is not editable in the
association_table, so when a user is created in the DB before they
are in the passwd file, they get a default UID of 65534 with no
way to edit it unless they go into a SQLite shell and manually edit
the column themselves.

Add "userid" as an editable field in the edit-user command.