view-user: raise error when user cannot be found
Problem: The view-user command still succeeds even when a user cannot
be found in the flux-accounting database.

Change the command to instead raise a ValueError when a user cannot be
found.

In flux-account-service.py, catch the exception, respond with an error,
and provide a message on why the command failed.