add_user: add check if association already active
Problem: add_user() only checks to see if an association has been
disabled before attempting to execute an INSERT statement on the
association_table, when it should also check to see if an
association currently exists *and* is active.

Add a check to add_user() to see if the association being addeed
already exists in the association_table; if they do, raise a
sqlite3.IntegrityError.