user_subcommands: raise error to caller
Problem: The functions called by all of the user subcommands handle any
errors raised during execution and just return a string containing the
error message. This behavior should be changed to be more Pythonic and
raise errors to the caller of the function.

Change the behavior of a number of the functions in user_subcommands.py
to raise any error to the caller of the function with a clear message
explaining what the error is. For sqlite3.OperationalError, provide
more context on what the error means and add a link to official docs
containing more information about OperationalError.