plugin: rm cur_run_jobs++ in job.state.priority
Problem: The multi-factor priority plugin currently increments
cur_run_jobs for a job when it enters job.state.priority. This is
incorrect because a job will not necessarily enter RUN state after it
receives a priority. Jobs enter SCHED state after receiving a priority
and could stay there while waiting for the requested resources before
actually running, or re-enter job.state.priority on a reprioritization
of all jobs after a flux-accounting update is sent to the plugin,
resulting in a double-increment of currently running jobs.

Remove the increment of a user/bank combo's current running jobs count
in priority_cb ().