plugin: add callback for job.state.run
Problem: The plugin needs a way to know when to increment the current
running jobs count for a user/bank combo now that it does not increment
the count when the job enters job.state.priority.

Add a new callback to the plugin to increment the current running jobs
count in job.state.run.