t: add sharness test for issue#262
Add a sharness test that tests the scenarios brought up in issue#262,
which revealed problems with how the plugin handled running and active
job counts. Add tests in this file that check for the following:

- Submit a job that is running and two more that are scheduled, but do
not run. Check that job counts are correct (1 running, 3 active).
- Run flux account-priority-update while the jobs are running/scheduled
and ensure that job counts are correct (1 running, 3 active).
- Change the priority of one of the scheduled jobs, reprioritize all
jobs, and ensure job counts are still correct (1 running, 3 active).
- Cancel one of the scheduled jobs and ensure job counts are still
correct (1 running, 2 active).
- Cancel the remaining running and scheduled jobs and ensure job counts
are still correct (0 running, 0 active).