bank_subcommands: raise error to caller
Problem: The functions called by the bank subcommands handle any
errors raised during execution and just return a string containing the
error message. This behavior should be changed to be more Pythonic and
raise errors to the caller of the function.

Change the behavior of a number of the functions in bank_subcommands.py
to raise any error to the caller of the function with a clear message
explaining what the error is.