t1007: edit tests to look for ValueError
Now that the *-bank commands raise errors to the caller of the
function, change the various sharness tests that look for these
errors to account for Python exceptions being raised during the
tests, i.e add "test_must_fail" before the commands that are
supposed to raise exceptions.

Adjust the expected output for a couple of the bank sharness tests to
match the error message that is raised during the tests.