service: make certain watchers accessible by all
Problem: All of the message handlers in the flux-accounting service are
only available to the instance owner, but some of them should be open
to all users. Commands that simply read from the flux-accounting DB,
like view-user, view-bank, etc. should be able to be run by any user
on the machine, while others like add-user, add-bank, should only be
able to be run by a privileged group of administrators.

Break up the "endpoints" list into two smaller lists, one that is
general and can be run by any user, and one that is privileged and is
only available to the instance owner. Specifically, when creating the
watchers from the general list, set a rolemask to FLUX_ROLE_USER when
creating the watcher.