plugin: clear bank_info->queues on update
Problem: THe "queues" field for a user/bank entry in the multi-priority
plugin does not get cleared on a flux account-priority-update, which
means if a user's accessible queues list gets cleared in the
flux-accounting database, the user can still submit jobs under those
queues.

Clear the "queues" list in the user's entry in the multi-factor priority
plugin on a flux account-priority-update.