plugin: clear queues map before an update
Problem: The 'queues' map in the plugin does not get cleared on an
update, so if queues are removed in the flux-accounting DB and then
an update is sent to the plugin via flux account-priority-update, the
old queues still persist in the plugin, which still prevents users
from being able to submit jobs to certain queues.

Clear the 'queues' map before receiving an update from the
flux-accounting DB to ensure consistency with the database, even if
queues are removed.