view-user: improve output of view-user command
Problem: The output of the view-user command is quite ugly now because
of how many columns there are in the association_table. Often times the
output wraps around the terminal which makes it hard to look at.

Change the format of the output of view-user to print user
information vertically instead of horizontally.