t1007: change tests that call grep on multiple str
Problem: The tests in t1007-flux-account.t that call "grep" on
multiple strings in a line of text are used incorrectly. It just greps
for the last string in the command, which isn't what we want. We want
to check for multiple strings in the output.

Change the way "grep" is used in the tests that use it to check for
multiple strings in output by passing the -w flag and concatenating
the various strings to look for with "|\."