plugin: reject jobs when user has no entry
Problem: The plugin currently holds jobs in PRIORITY when a user
submits a job but does not have an entry in the flux-accounting DB,
regardless if the plugin has *any* data about users/banks. This can
be confusing to users who are submitting jobs on machines for the first
time.

Change the behavior of the job.validate and job.new callbacks to check
the state of the plugin's internal data before deciding what to do with
a newly submitted job:

If the plugin has *no* data about any users or banks, the plugin will
hold the job in PRIORITY state until it receives some data.

If the plugin has *some* data about any users or banks and the
submitted job is from a user who does not yet have an entry in the
plugin's internal map, the job will be a rejected with a message saying
"no bank found for userid: <userid>."