t: add test for jobs in different plugin states
Add a new sharness test file that tests submitting a job while the
plugin has no user/bank data and ensuring that the job is held in
PRIORITY until it receives some data. It then submits a job while
the plugin is loaded with user/bank data and ensures that jobs
submitted by users who do not have a user/bank entry in the plugin
are rejected.