bank_subcommands: improve view-bank "-t" option
Problem: There is no way to see an overarching view of a
flux-accounting hierarchy without running "flux account-shares." It
would be nice if you could see this view with a Python command.

Edit the "-t" option of the view-bank to more closely match the output
of the "flux account-shares" command in Python, which includes job
usage values for banks and fair share values for users all in one
place.