docker/README.md: remove references to bionic
Problem: There are a couple of references to the bionic container in the
README for the docker/ directory. Bionic is being removed in favor of
jammy.

Remove the references to bionic and change them to jammy. Remove the
statement about multiple versions of Python being able to be tested with
the bionic container.