plugin: add bank info lookup helper function
Problem: There are multiple spots in the plugin code where we need to
look up user/bank information, and the code isn't super easy to read
because a map iterator needs to be declared and used to look up the
information in a map. It might be easier to read and understand what is
going on if this lookup was moved to a helper function.

Create a helper function, get_bank_info (), that accepts a bank_info
map struct iterator, userid, and a bank parameter and performs this
lookup in the map. It returns a pair: a return code as well as the
iterator to the user/bank map.

In the spots where this new helper function is called, the same logic
is retained that handles validating the user/bank information passed
in, i.e rejecting the job or raising a job exception because the
user/bank information passed in cannot be validated.