python: rm empty quotechar arg from csv.writer
Problem: The csv.writer object created in job_archive_interface.py uses
an optional arg quotechar that is initialized to an empty string (""),
but the docs state that this must be at least a one-character string,
causing some tests to fail. I had copied the initialization of the
csv.writer object from somewhere when I first wrote this file,
evidently not using it correctly.

Just remove this optional argument from the object initialization.