t: add valgrind testing
Problem: flux-accounting currently has no way to run tests under
valgrind, but it is useful to run tests in this mode when working with
the multi-factor priority plugin.

Copy the t/valgrind/ folder from flux-core and add it to
flux-accounting.

Copy over the t5000-valgrind.t file from flux-core and add it to
flux-accounting's test suite.

Add a simple flux-accounting-specific valgrind test that just loads the
multi-factor priority plugin and submits a job.