ci: update github actions
Problem: flux-accounting's GitHub actions uses a number of deprecated
versions of actions that need updating. Running CI in flux-accounting
reports a number of different warnings.

Update actions/checkout to v3.

Update actions/setup-python to v4 which matches flux-core.

Drop the use of the set-output command, which is deprecated.

Remove the prep-release action from flux-accounting's workflow file,
which is probably no longer needed.

Use softprops/action-gh-release for the create-release action, which is
much simpler to use for creating flux-accounting releases.