plugin: add extra check for contents of users map
Problem: In validate_cb (), the multi-factor priority plugin has a bug
where even if it only has the special does-not-exist ("DNE") entries for
at least one user/bank (indicating that the plugin has not received any
flux-accounting data yet), subsequently submitted jobs by other users
get rejected with the "no bank found for user" error message, which is
the rejection message typically output when the plugin has valid
user/bank data.

Add a helper function to the priority plugin that scans the users map
and checks the default bank for each user. As soon as it finds a
user/bank with a default bank that is not the special "DNE" entry,
return false. Otherwise, return true.

Use this helper function in the callback for job.validate to ensure that
the plugin's internal user/bank map contains valid user/bank data before
rejecting a user's job because they don't have a valid entry in the
flux-accounting database.