plugin: move queue priority assignment to job.new
Problem: The multi-factor priority plugin assigns the priority
associated with a given queue to the bank_info struct associated with
each job in job.validate. Now that Flux can accept job updates and jobs
can re-enter job.validate multiple times, this assignment should be
moved to job.new, and job.validate should only be used for validating a
queue if one is passed in.

Remove the assignment of the priority associated with a passed in queue
in job.validate.

In job.new, instead of only assigning a queue's priority in the case
where the bank_info struct unpacked is NULL, move the assignment to
later in the callback so that it is executed in all cases.

This will result in the bank_info struct associated with each job to
have an associated queue priority by the end of job.new.