plugin: add function to retrieve user/bank info
Create a helper function, get_bank_info (), that accepts a userid and a
bank parameter and performs a lookup in the plugin's internal users map.
It returns a pair: a return code as well as the iterator to the
user/bank map that contains the information to be associated with each
job's bank_info struct, such as their job limits and permissible
queues.