plugin: add job-update callback
Problem: Flux can now accept job updates, but the multi-factor priority
plugin does not have a callback to look at these updates, particularly
for attributes that it is involved in validating.

Add a job-update callback to the plugin. Unpack the userid, bank, and
queue attributes from jobspec and assign the priority associated with
the new queue if one is passed in to the bank_info struct that is
associated with the job.