t: add tests for updating queue
Problem: flux-accounting has no tests that update the queue of a pending
job and ensures that the priority plugin can handle cases where the job
should be updated and where the job should not be updated because the
queue cannot be validated for the user/bank's job.

Add some tests.