t: change check for job states
Problem: There are some tests in the flux-accounting testsuite that
might produce a race condition, especially when run under valgrind.
Some of the tests look at specific job states, but there is a chance
the job may not have reached a certain state, like DEPEND, by the time
flux job info is run.

In the testsuite, change the way that specific job states are checked
by using "flux job wait-event".