t1011: remove check for specific job usage value
Problem: t1011-job-archive-interface.t checks for a specific job usage
value after running flux account update-usage, but there is a chance
that the time elapsed for the submitted jobs is longer than what is
assumed in the testsuite.

Just remove the tests that check for a specific job usage value.