t1019: change order of job cancellations
Problem: Running t1019-mf-priority-info-fetch.t under valgrind causes
the third submitted job to enter an alloc event by the time the first
two are canceled, which we don't want for the purpose of the tests
because we are checking for certain cur_run_jobs and cur_active_jobs
values when we query the plugin's internal map.

Change the order of the job cancellations to cancel the last submitted
job first so that it never enters the alloc event while the first two
are in the process of being cancelled.