view-bank: fix -t option for sub bank w/ users
Problem: the -t option of the view-bank command does not print out a
full hierarchy for a sub bank with users in it because there is an "if"
check that looks to see if the bank has any sub banks in it before
calling the recursive function that prints out any potential information
under it, but it won't call this recursive function if it doesn't have
any sub banks.

Remove the "if" check before calling the recursive function so that it
is always called whether a sub bank has any additional sub banks, users,
or neither.