plugin: add update_queue_cb ()
Problem: The multi-factor priority plugin uses the same callback for
validating an updated queue and for actually updating the queue on a
user/bank's job. This will be an issue when additional "updateable"
attributes are added to the plugin, such as banks.

Add a new callback to the plugin, update_queue_cb (), that is
responsible for validating an updated queue and rejects the update
if it is not valid for the user/bank. Assign the callback for the
"job.update.attributes.system.queue" topic string.

As a result of adding the new update_queue_cb () callback, change the
job_updated () callback to unpack the "updates" JSON object when
unpacking arguments and see if there is an update for the queue
attribute. Check its contents for NULL before applying its update to
the job that requested the queue update.

Add a comment at the top of the job_updated () function with a
description of what it does.