plugin: add job.update for banks
Problem: flux-accounting currently has no support for a user to update
the bank of their job.

Add a new callback to the plugin which looks for the
"job.update.attributes.system.bank" topic string and validates the new
bank that the user is trying to update their job to. If they have access
to the bank they want to update their job to but it is currently at its
max active jobs OR max running jobs limits, reject the update.

Add a check for this update in the job.update callback which will look
for an updated bank. If the updated bank has been validated, change the
bank_info struct assigned to that job to reflect the attributes of the
new bank.