validate_cb (): remove unused/unneeded variables
Problem: There are a number of variables in validate_cb () that are
either unused or unneeded because the things they were originally used
for were moved to other callbacks in the priority plugin. The
queue_factor variable in particular can be removed in favor of just
using the function call in its place.

Remove the unused/unneeded variables from the validate_cb () callback.