validate_cb (): use lookup function for user/bank
Problem: There is a new lookup helper function in mf_priority.cpp that
can look up user/bank information with the userid and an optional bank,
but the job.validate callback does not make use of it.

Convert the user/bank lookup in validate_cb () to use this new lookup
helper function.