plugin: create new user_bank_info class
Problem: The plugin uses its own bank_info struct to hold user/bank
information and has a number of methods for access and modification
of these structs. As the plugin's feature set has grown, so have the
requirements for the bank_info struct, resulting in a very large and
hard-to-parse piece code.

Begin to clean up this plugin. Start by creating a new user_bank_info
class and place it in a separate file that gets compiled with the
plugin. Replace all instances of "struct bank_info" with the new
"user_bank_info" class type.