validate_cb (): use new helper function
Problem: There are now helper functions available for the plugin to make
use of to help with lookups and access of the internal map for users and
banks, but the plugin does not make use of them.

Restructure the validate_cb () function to make use of the new
get_user_info () helper function.

Change the error message for when a user/bank cannot be found and adjust
the respective tests that look for this error message.

Improve the comments in this function/add new ones to help with some of
the clarity in this function.