plugin: keep jobs in PRIORITY after reprioritize
Problem: the priority plugin will raise an exception on a job if it is
held in SCHED state while the plugin is reloaded (or Flux is restarted)
and jobs are reprioritized without first loading flux-accounting data to
this plugin. This behavior is not graceful and we should instead
continue to hold a job in PRIORITY while the plugin waits to receive
flux-accounting data.

Add a check of the plugin's internal map to see if we are still waiting
on flux-accounting data to be loaded in; if so, continue to hold the job
while we wait for data.

Add a sharness test that reproduces the issue raised in #406 and ensure
that jobs continue to be held after a reprioritization without loading
flux-accounting data to the priority plugin.