plugin: move check_map_for_dne_only ()
Problem: The helper function check_map_for_dne_only () is specific to
the internal maps that store user/bank information in the priority
plugin. It might make more sense to include this in the external files
that are centered around user/bank information instead.

Move check_map_for_dne_only () to bank_info.cpp. Add both maps that
store user/bank information as arguments to the function. Edit the
calls to this function in validate_cb () and priority_cb () to
include both maps as arguments.