association: add new lookup function
Problem: There is some functionality in the priority plugin
(mf_priority.cpp) that is specific to accessing Association objects
in a map data structure. These methods can be abstracted out to the
newly created association.* files to clean up some of the plugin code.

Add a new function to the association.cpp that handles looking up an
Association in the users map, or returning nullptr on failure.