accounting: add to_json method, function
Problem: the accounting information that is stored in the plugin's
internal map cannot be retrieved and printed out in a readable format
like JSON.

Add a method to the Association class called to_json (), which will
convert an Association object to JSON format.

Add a helper function to accounting.cpp called convert_map_to_json (),
which will convert all of the users/banks in the plugin's internal users
map to JSON format and return it as a pointer to a json_t.

Add jansson as a library in src/Makefile.am.