plugin: use new external function in query_cb ()
Problem: The priority plugin uses multiple helper functions defined in
the plugin itself that convert the user/bank data in the users map to
JSON format to be returned in the plugin.query callback, but there are
now external functions that do this much more cleanly.

Remove the helper functions defined in the plugin itself and instead
use the external functions defined in association.cpp to convert the
accounting data to JSON format.