t1019: update expected output of flux jobtap query
Problem: The new external functions that convert the accounting data in
the plugin's internal map has been updated and now returns all info per
user/bank in the map, but the tests in t1019 check for JSON output that
matches the old output.

Update the expected output to match the new updated JSON objects that
are constructed for each user/bank in the plugin.