t1019: add queue to one of the users for testing
Problem: None of the expected output of the tests in t1019 have queues
in them because no associations belong to any queues when their info is
sent to the priority plugin.

Add some queues to the flux-accounting DB and add a queue to one of the
users so that we can test expected output for the "queues" key-value
pair in the expected output.