plugin: expand split_string () and move it
Problem: The split_string () helper function can only accept a list of
queues to parse when unpacking user/bank information from the
flux-accounting DB, but there will also exist a need in the future to
be able to parse other string lists as well. Since both lists can be
parsed the same way, this helper function should be expanded. Also,
this helper function is specific to parsing flux-accounting data. It
might make more sense to instead place this in accounting.cpp.

Move the function definition out of the priority plugin code and
into accounting.cpp.

Expand the split_string () helper function to accept a general string
list of items and place them in one of the vector members of an
Association object. Rename it to split_string_and_push_back ().