plugin: change variable name in rec_update_cb ()
Problem: The variable name that's assigned to the unpacked list of
queues in rec_update_cb () is also named queues, which conflicts with
the global map variable "queues", which causes confusion as to what is
being passed as arguments to the split_string_and_push_back () function.

Rename the variable to "assoc_queues" to indicate that it represents the
list of queues that the association belongs to.