plugin: remove definition of split_string ()
Problem: The priority plugin defines a function to split a string of
queues, but accounting.cpp now has a more general function to accomplish
this.

Remove the definition for split_string () from the priority plugin and
use the newly defined split_string_and_push_back () function defined in
accounting.cpp.