plugin: improve job.new callback
Problem: The callback for job.new still performs a manual lookup of an
Association, but there now exists external functions defined in
accounting.cpp that the callback should make use of. It is also lacking
a function description and has some leftover variables that are
unutilized.

Clean up this function by using the new external functions that perform
lookups of an association. Remove unused variables from the function.
Add a function description at the top.