update_queue_cb/job_updated: update variable names
Problem: The variable name chosen for an updated queue is just named
"queue"; this is poorly named and might cause confusion for what it is
actually unpacking, which is an updated queue. Also, the variable that
holds the Association object in job_updated () is inconsistently named
"b" instead of "a" like the other callbacks that unpack an Association
object.

Update "queue" to "updated_queue". Update "Association *b" to
"Association *a" in job_updated ().