plugin: remove get_bank_info (), typedef, enums
Problem: the plugin defined its own helper function for looking up
associations in its internal map, but it was clunky, and now
accounting.cpp has a much cleaner implementation that is used within
the plugin code.

Remove this helper function and its associated typedef and enums since
it is no longer needed in the plugin code.