plugin: improve job.state.priority callback
Problem: The callback for job.state.priority uses a manual lookup for
associations in the case where we need to re-look up a user's accounting
information, but accounting.cpp has a cleanly defined implementation of
this now.

Reformat the job.state.priority callback to make use of the new external
lookup function defined in job.state.priority. Improve some of the
comments throughout the function and add new ones. Adjust the test that
checks for an exception message in job.state.priority to account for the
new message.