accounting.cpp: improve get_queue_info ()
Problem: The get_queue_info () function could use some minor
improvements to its comment structure and some clean up to its if-else
checks.

Slightly modify some of the comments in this function and use auto for
the declaration of the vector iterators used in get_queue_info ().
Clean up some of the if-else checks where "else" is not necessarily
needed.