plugin: change struct queue_info -> class Queue
Problme: The plugin used to define a struct for queue_info objects, but
now that we have external files for flux-accounting-specific stuff, it
would make more sense to instead define a class for Queue objects that
is structured similarly to the Association class.

Convert the struct queue_info definition to a class Queue definition.
Replace the usage of struct queue_info in both get_queue_info () and in
the plugin code itself with the new class Queue.

Use const references as parameters in the get_queue_info () function
definition.